cyclicReferences folder: Remove describe clause "cyclic references" and rename files
